move body calculation to group level
