+++ app/components/demonstrator-group-data.js
@@ -8,55 +8,8 @@ export default Ember.Component.extend({
   session: Ember.inject.service('session'),
   classNames: ['demonstrator-group'],
   header: ['Név', 'Neptun', "Csoport", 'Feladattípus', 'Feltöltés ideje', 'Beugró érdemjegy', 'Beadandó érdemjegy', 'Labor érdemjegy', 'Pót'],
-    rowIndecies: ['StudentRegistration.User.displayName', 'StudentRegistration.User.neptun', 'StudentRegistration.neptunCourseCode', 'ExerciseSheet.ExerciseType.shortName', 'firstCorrectableDeliverable.formattedLastSubmittedDate', 'firstEntryTest.grade', 'firstCorrectableDeliverable.grade', 'grade', 'supplementary'],
+  rowIndecies: ['StudentRegistration.User.displayName', 'StudentRegistration.User.neptun', 'StudentRegistration.neptunCourseCode', 'ExerciseSheet.ExerciseType.shortName', 'firstCorrectableDeliverable.formattedLastSubmittedDate', 'firstEntryTest.grade', 'firstCorrectableDeliverable.grade', 'grade', 'supplementary'],
 
-  body: Ember.computed('currentEventTemplate', function () {
-    return new RSVP.Promise((resolve, reject) => {
-      if (this.get('currentEventTemplate.id')) {
-        this.get('store').query('event', {
-          filter: {
-            eventTemplateId: this.get('currentEventTemplate.id')
-          }
-        }).then(events => {
-          let body = [];
-          resolve(events
-            .map(event => {
-              event.set('formattedDate', dateformat([event.get('date')]));
-              return event;
-            })
-            .sort((lhs, rhs) => {
-              const lhsName = lhs.get('StudentRegistration.User.displayName'), rhsName = rhs.get('StudentRegistration.User.displayName');
-              const lhsAttempt = lhs.get('attempt'), rhsAttempt = rhs.get('attempt');
-
-              if (lhsAttempt > rhsAttempt) {
-                return -1;
-              }
-
-              if (lhsAttempt < rhsAttempt) {
-                return 1;
-              }
-
-              if (lhsAttempt === rhsAttempt) {
-                if (lhsName < rhsName) {
-                  return -1;
-                }
-                if (lhsName > rhsName) {
-                  return 1;
-                }
-              }
-              return 0;
-            })
-          );
-        }, err => {
-          console.error(err);
-          reject(err);
-        });
-      }
-      else {
-        resolve([]);
-      }
-    });
-  }),
 
   download(supplementary) {
     const form = document.createElement('form');
