+++ app/templates/components/demonstrator-group-data.hbs
@@ -12,4 +12,4 @@
 </div>
 <a class="button" {{action 'generateSheet'}}>generálás</a>
 <a class="button no-margin-left" {{action 'generateSheetSupplementary'}}>generálás pót</a>
-{{table-component rowIndecies=rowIndecies header=header body=(await body) addClass='demonstrator-table' onClickRowAction='evaluateEvent' clickable=true evaluateEvent=(action "evaluateEvent")}}
+{{table-component rowIndecies=rowIndecies header=header body=(await events) addClass='demonstrator-table' onClickRowAction='evaluateEvent' clickable=true evaluateEvent=(action "evaluateEvent")}}
