+++ app/templates/components/demonstrator-group.hbs
@@ -1,1 +1,1 @@
-{{demonstrator-group-data currentEventTemplate=currentEventTemplate evaluateEvent=(action 'evaluateEvent')}}
+{{demonstrator-group-data currentEventTemplate=currentEventTemplate events=events evaluateEvent=(action 'evaluateEvent')}}
